+++ lib/lightning_web/live/project_live/collaborators.ex
@@ -29,7 +29,7 @@ defmodule LightningWeb.ProjectLive.Collaborators do
     schema
     |> cast(attrs, [:email, :role])
     |> validate_required([:email, :role])
-    |> validate_format(:email, ~r/^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/)
+    |> validate_format(:email, ~r/^[\w-\.]+@([\w-]+\.)+[\w-]{2,}$/)
   end
 
   @spec prepare_for_insertion(%__MODULE__{}, map(), [map(), ...]) ::