+++ lib/lightning_web/live/workflow_live/dashboard_components.ex
@@ -352,7 +352,8 @@ defmodule LightningWeb.WorkflowLive.DashboardComponents do
         disabled={@disabled}
         tooltip={@tooltip}
         phx-click={
-          if !@disabled, do: JS.navigate(~p"/projects/#{@project_id}/w/new")
+          if !@disabled,
+            do: JS.navigate(~p"/projects/#{@project_id}/w/new")
         }
         class="col-span-1 w-full"
         role="button"